+++ R/backend-secret-service.R
@@ -26,7 +26,7 @@
 #' \dontrun{
 #' ## This only works on Linux, typically desktop Linux
 #' kb <- backend_secret_service$new()
-#' kb$create_keyring("foobar")
+#' kb$keyring_create("foobar")
 #' kb$set_default_keyring("foobar")
 #' kb$set_with_value("service", password = "secret")
 #' kb$get("service")
@@ -45,7 +45,7 @@ backend_secret_service <- R6Class(
     get = function(service, username = NULL, keyring = NULL)
       b_ss_get(self, private, service, username, keyring),
     get_raw = function(service, username = NULL, keyring = NULL)
-      b_ss_get_raw(self, private, service, username = NULL, keyring = NULL),
+      b_ss_get_raw(self, private, service, username, keyring),
     set = function(service, username = NULL, keyring = NULL)
       b_ss_set(self, private, service, username, keyring),
     set_with_value = function(service, username = NULL, password = NULL,
@@ -101,15 +101,22 @@ b_ss_init <- function(self, private, keyring) {
 }
 
 b_ss_get <- function(self, private, service, username, keyring) {
-  keyring <- keyring %||% private$keyring
-  res <- .Call("keyring_secret_service_get", keyring, service, username)
+  res <- b_ss_get_raw(self, private, service, username, keyring)
   if (any(res == 0)) {
     stop("Key contains embedded null bytes, use get_raw()")
   }
   rawToChar(res)
 }
 
+b_ss_get_raw <- function(self, private, service, username, keyring) {
+  username <- username %||% getOption("keyring_username")
+  keyring <- keyring %||% private$keyring
+  res <- .Call("keyring_secret_service_get", keyring, service, username)
+  res
+}
+
 b_ss_set <- function(self, private, service, username, keyring) {
+  username <- username %||% getOption("keyring_username")
   password <- get_pass()
   b_ss_set_with_value(self, private, service, username, password, keyring)
   invisible(self)
@@ -117,6 +124,7 @@ b_ss_set <- function(self, private, service, username, keyring) {
 
 b_ss_set_with_value <- function(self, private, service, username, password,
                                 keyring) {
+  username <- username %||% getOption("keyring_username")
   keyring <- keyring %||% private$keyring
   .Call("keyring_secret_service_set", keyring, service, username,
         charToRaw(password))
@@ -125,6 +133,7 @@ b_ss_set_with_value <- function(self, private, service, username, password,
 
 b_ss_set_with_raw_value <- function(self, private, service, username, password,
                                     keyring) {
+  username <- username %||% getOption("keyring_username")
   keyring <- keyring %||% private$keyring
   .Call("keyring_secret_service_set", keyring, service, username,
         password)
@@ -132,6 +141,7 @@ b_ss_set_with_raw_value <- function(self, private, service, username, password,
 }
 
 b_ss_delete <- function(self, private, service, username, keyring) {
+  username <- username %||% getOption("keyring_username")
   keyring <- keyring %||% private$keyring
   .Call("keyring_secret_service_delete", keyring, service, username)
   invisible(self)